make Indel.Distance disposable fix PartialRatioStrategyT.cs to use fixed algorithm code cleanup
